moving Location & Tag components to /draftPost
