+++ frontend/src/components/draftPost/Location.tsx
@@ -1,6 +1,6 @@
 import { MapPin, X } from 'lucide-react';
-import { Button } from './ui/button';
-import { Input } from './ui/input';
+import { Button } from '../ui/button';
+import { Input } from '../ui/input';
 
 export const PostLocation = ({
   locationName,
